migrate to slick 3.2.1, scala 1.12

+++ src/main/scala/ch/seidel/kutu/data/ResourceExchanger.scala
@@ -122,9 +122,9 @@ object ResourceExchanger extends KutuService {
           case Some(w) => w.id
           case None => wettkampfid
         },
-        noteD = BigDecimal.valueOf(fields(wertungenHeader("noteD"))),
-        noteE = BigDecimal.valueOf(fields(wertungenHeader("noteE"))),
-        endnote = BigDecimal.valueOf(fields(wertungenHeader("endnote"))),
+        noteD = fields(wertungenHeader("noteD")),
+        noteE = fields(wertungenHeader("noteE")),
+        endnote = fields(wertungenHeader("endnote")),
         riege = if(fields(wertungenHeader("riege")).length > 0) Some(fields(wertungenHeader("riege"))) else None,
         riege2 = if(fields(wertungenHeader("riege2")).length > 0) Some(fields(wertungenHeader("riege2"))) else None
       )
